flask-celery-example/requirements.txt
amqp==2.1.4
appdirs==1.4.3
backports-abc==0.5
backports.ssl-match-hostname==3.5.0.1
billiard==3.5.0.2
blinker==1.4
celery==4.0.2
certifi==2017.1.23
click==6.7
Flask==0.12
Flask-Mail==0.9.1
itsdangerous==0.24
Jinja2==2.9.5
kombu==4.0.2
MarkupSafe==1.0
packaging==16.8
pkg-resources==0.0.0
pyparsing==2.2.0
pytz==2016.10
redis==2.10.5
singledispatch==3.4.0.3
six==1.10.0
tornado==4.4.2
vine==1.1.3
websocket-client==0.40.0
Werkzeug==0.11.15
